Carrier index
The initial implementation of a carrier inventory was in the organization#show action, which (1) does not feel right to me and (2) causes problems if we want to make the inventory the root page. Which organization should be displayed? We want to put off implementing organizations until we in fact have more than one organization
Looking at the commit history for these two files, it seems I am only changing an index that was there as a placeholder and returning the organization#show action to the version it was in before inventory made its way in there