Add links for users to select specific carriers
When users are viewing the list of carriers, there are currently no
links on any of the cards for them to view more details about the
carrier they are interested in. This change adds links so the name of
the carrier along with the photo are clickable.

This also resolves a broken feature test which expects a link with the
content of a carrier name to exist.