Fix flaky carrier specs
We have had multiple times where builds would fail, mostly on `carrier.location.name` with a `undefined method name for nil`. Seems like the reason for it is that location fixtures were not enabled when the carrier fixtures were. I do not like this fix (nor do I like fixtures) but this should resolve our issues for the time being