FIX: Security hole and failing spec (#180)
PR #176 introduced different message when login is not successful.
But not all Devise locales are changed so we had situation depending on
what is invalid `email` or `password` we will get different message.
Which exposes security risk, because someone can figure out valid email
more easily.

Also, because of that specs are failing. So I've fixed the local and
expected message in the specs.