Resolves #182
I'm pushing the rubocop + rubocop-rspec additions to the Gemfile + a customized .rubocop.yml file that should adapt nicely to the current state of the project. The reason why I added the custom rules is not to make a ton of disruptions on the project.

The project uses a mix of single and double qoutes, line lengths usually don't go over 120 chars, the top level classes are not documentated etc, which is all okay, and I think the bigger part is not having trailing whitespaces and inconsistent indentation.

If everyone's alright with this setup, i'd merge it in and then do a fast followup with rubocops autocorrect feature.