Introduce a basic carrier checkout ability.
This PR adds basic carrier checkout ability.
If the user is admin or volunteer he/she will see `Checkout` button on
the Carrier.

When `Checkout` button is clicked modal will open to find the member and
populate `due_date`. Only users with role member are shown inside the
select and `default_date` from the carrier is automatically selected on
the forms `due_date`.

`checkout_volunteer` is set as user who is logged-in.

NOTE:
I didn't spend to much time on the styling, some of the styles are
intentionally inlined, since styles are gonna be covered in another PR.