Drop `rolify` gem
I don't think we need `rolify` gem for the following reasons

1. We don't have any advanced roles in the system, just basic roles on
the user.
2. New developers are having a hard time to figure out how `rolify` works
3. Since we don't have complex roles, we don't need multiple tables for
roles(rolify implementation) and hence alyways executing unecessary db
queries.
4. In our use case we don't need new gem to maintain.

===============================================================

Since our roles are simple atm, this PR removes `rolify` and adds simple
`role` column on the users which is powered by `Rails enum`.

If in the future we feel need to introduce `roles` table it could be
done easily inside the migration.